[discover] update advanced settings to restore state completely (#6969)
* advanced settings toggle restore language switcher


* clean up toggle logic to settings


---------
